refactor: ReservationRepository 인터페이스로 변경 및 JdbcReservationRepository 추가
