+++ src/main/java/roomescape/reservation/ReservationRepository.java
@@ -1,34 +1,12 @@
 package roomescape.reservation;
 
-import java.util.Optional;
-import org.springframework.stereotype.Repository;
 import java.util.List;
+import java.util.Optional;
 
-@Repository
-public class ReservationRepository {
-    private final ReservationDao reservationDao;
-
-    public ReservationRepository(ReservationDao reservationDao) {
-        this.reservationDao = reservationDao;
-    }
-
-    public List<Reservation> findAll() {
-        return reservationDao.findAll();
-    }
-
-    public Reservation save(Reservation reservation) {
-        return reservationDao.save(reservation);
-    }
-
-    public void delete(long id) {
-        reservationDao.delete(id);
-    }
-
-    public int countByTimeId(long timeId) {
-        return reservationDao.countByTimeId(timeId);
-    }
-
-    public Optional<Reservation> findById(long id) {
-        return reservationDao.findById(id);
-    }
+public interface ReservationRepository {
+    List<Reservation> findAll();
+    Reservation save(Reservation reservation);
+    void delete(long id);
+    int countByTimeId(long timeId);
+    Optional<Reservation> findById(long id);
 }
